+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#2132](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2132))
 - `opentelemetry-resource-detector-azure` Changed timeout to 4 seconds due to [timeout bug](https://github.com/open-telemetry/opentelemetry-python/issues/3644)
   ([#2136](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2136))
+- Remove erroneous `type` attribute from `opentelemetry-instrumentation-asgi`
+  ([#2145](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/2145))
 
 ## Version 1.22.0/0.43b0 (2023-12-14)